+++ lib/features/home/presentation/model/cv_result_model.dart
@@ -1,5 +1,5 @@
 class CVResultModel {
-  int? time;
+  double? time;
   String? bodyLanguageClass;
 
   CVResultModel({this.time, this.bodyLanguageClass});
